JAMES-2904 Move ResilientClusterProvider to backend
